+++ DietHelper.Common/Models/MealEntries/UserMealEntryIngredient.cs
@@ -8,7 +8,7 @@ namespace DietHelper.Common.Models.MealEntries
         public int Id { get; set; }
         public int UserMealEntryId { get; set; }
         public int UserProductId { get; set; }
-        public decimal Quantity { get; set; }
+        public double Quantity { get; set; }
         public string ProductNameSnapshot { get; set; } = string.Empty;
         public NutritionInfo ProductNutritionInfoSnapshot { get; set; } = new();
         public bool IsDeleted { get; set; }
